:tada: Update to version 1.8 (build 20211002)
This is no rewrite from scratch, but changes a lot of the script's behavior. The scripts now run unprivileged and store files in '/usr/local/acme'. The 'account.key' is now expected in '/etc/acme', as well as a new 'config.env' file that allows some basic configuration (like using another CA and setting a contact email address). Instead of symlinking the latest certs to the 'live' directory, `acme-issue` will now copy the files (this improves compatiblity with ionotify). The renewal script can now retry if a renewal failed earlier. Last but not least, the documentation was rewritten from scratch (there's a 'README.md' now) and some parts of the scripts were streamlined.

There's an unreleased version 1.7 in between we're skipping now. Version 1.7 was used internally only and just for a couple of months. It was pretty much tied to a single setup, so it's not really worth releasing this.