Verify all certificates after issuance
Pass the '--no-verify' option to disable the checks. `openssl verify` will not just verify that all certs of the chain are valid X.509 certs (as we did before and still do), but will perform stricter compliance and policy checks, will check that the certs match their supposed types and common name (CN), and will check that all certs form a complete and trusted chain. The latter requires the root cert to be installed in the trust store of the system.